Abstract out help window and internships window methods
