+++ src/main/java/seedu/address/logic/commands/util/CommandUtil.java
@@ -100,8 +100,6 @@ public abstract class CommandUtil {
         }
     }
 
-
-
     /**
      * Checks which tab's display needs to be switch.
      *
